use common raw implementations, convert memory cache to store Vec<u8> keys/values to facilitate this
